+++ app/components/DetailPanel.tsx
@@ -13,7 +13,7 @@ import { themed } from "../theme/theme"
 import { TimelineItem } from "../types"
 import { TreeViewWithProvider } from "./TreeView"
 import ActionButton from "./ActionButton"
-import { Tooltip } from "./Tooltip/Tooltip"
+import { Tooltip } from "./Tooltip"
 import IRClipboard from "../native/IRClipboard/NativeIRClipboard"
 import { $flex } from "../theme/basics"
 import { formatTime } from "../utils/formatTime"
